# apps/gkx-web-app/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/exercises": {
      "post": {
        "operationId": "ExercisesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exercise created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Exercise created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ExerciseModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Crear ejercicio",
        "tags": [
          "Exercises"
        ]
      },
      "get": {
        "operationId": "ExercisesController_findAll",
        "parameters": [
          {
            "name": "trainingContentId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercises retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Exercises retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ExerciseModel"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Listar ejercicios",
        "tags": [
          "Exercises"
        ]
      }
    },
    "/api/exercises/{id}": {
      "get": {
        "operationId": "ExercisesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador del ejercicio",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Exercise retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ExerciseModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Obtener ejercicio por id",
        "tags": [
          "Exercises"
        ]
      },
      "patch": {
        "operationId": "ExercisesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador del ejercicio",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercise updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Exercise updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ExerciseModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Actualizar ejercicio",
        "tags": [
          "Exercises"
        ]
      },
      "delete": {
        "operationId": "ExercisesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador del ejercicio",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Exercise deleted successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeleteDataModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Eliminar ejercicio",
        "tags": [
          "Exercises"
        ]
      }
    },
...
    "/api/training-sessions/{sessionId}/exercises/{sessionExerciseId}": {
      "patch": {
        "operationId": "SessionExercisesNestedController_update",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Identificador de la sesion",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sessionExerciseId",
            "required": true,
            "in": "path",
            "description": "Identificador de session-exercise",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionExerciseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session exercise updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session exercise updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SessionExerciseModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Actualizar ejercicio de una sesion",
        "tags": [
          "Session Exercises"
        ]
      },
      "delete": {
        "operationId": "SessionExercisesNestedController_remove",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Identificador de la sesion",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sessionExerciseId",
            "required": true,
            "in": "path",
            "description": "Identificador de session-exercise",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session exercise deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session exercise deleted successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeleteDataModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Eliminar ejercicio de una sesion",
        "tags": [
          "Session Exercises"
        ]
      }
    },
...
  },
  "info": {
    "title": "GKX API",
    "description": "Documentacion de la API GKX",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Ingresa el token JWT en formato Bearer"
      }
    },
    "schemas": {
...
      "DeleteDataModel": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "ExerciseModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "trainingContentId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "instructions": {
            "type": "object",
            "nullable": true
          },
          "objective": {
            "type": "object",
            "nullable": true
          },
          "durationMinutes": {
            "type": "object",
            "nullable": true
          },
          "repetitions": {
            "type": "object",
            "nullable": true
          },
          "restSeconds": {
            "type": "object",
            "nullable": true
          },
          "equipment": {
            "type": "object",
            "nullable": true
          },
          "videoUrl": {
            "type": "object",
            "nullable": true
          },
          "difficulty": {
            "type": "object",
            "nullable": true
          },
          "order": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "trainingContentId",
          "name",
          "description",
          "instructions",
          "objective",
          "durationMinutes",
          "repetitions",
          "restSeconds",
          "equipment",
          "videoUrl",
          "difficulty",
          "order",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateExerciseDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "trainingContentId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Reaccion a estimulo visual"
          },
          "description": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "durationMinutes": {
            "type": "number",
            "example": 5
          },
          "repetitions": {
            "type": "number",
            "example": 6
          },
          "restSeconds": {
            "type": "number",
            "example": 30
          },
          "equipment": {
            "type": "string"
          },
          "videoUrl": {
            "type": "string",
            "example": "https://video.example.com/clip"
          },
          "difficulty": {
            "type": "string",
            "example": "MEDIUM"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        },
        "required": [
          "tenantId",
          "trainingContentId",
          "name"
        ]
      },
      "UpdateExerciseDto": {
        "type": "object",
        "properties": {}
      },
...
      "SessionExerciseModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionContentId": {
            "type": "string",
            "format": "uuid"
          },
          "exerciseId": {
            "type": "string",
            "format": "uuid"
          },
          "order": {
            "type": "number"
          },
          "selected": {
            "type": "boolean"
          },
          "customDurationMinutes": {
            "type": "object",
            "nullable": true
          },
          "customRepetitions": {
            "type": "object",
            "nullable": true
          },
          "customRestSeconds": {
            "type": "object",
            "nullable": true
          },
          "coachNotes": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "sessionId",
          "sessionContentId",
          "exerciseId",
          "order",
          "selected",
          "customDurationMinutes",
          "customRepetitions",
          "customRestSeconds",
          "coachNotes",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateSessionExerciseDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}